+++ plugins/module_utils/retries.py
@@ -29,6 +29,8 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
+from functools import wraps
+
 try:
     from botocore.exceptions import ClientError
     HAS_BOTO3 = True
@@ -76,3 +78,36 @@ class AWSRetry(CloudRetry):
             retry_on.extend(catch_extra_error_codes)
 
         return response_code in retry_on
+
+
+class RetryingBotoClientWrapper(object):
+    __never_wait = (
+        'get_paginator', 'can_paginate',
+        'get_waiter', 'generate_presigned_url',
+    )
+
+    def __init__(self, client, retry):
+        self.client = client
+        self.retry = retry
+
+    def _create_optional_retry_wrapper_function(self, unwrapped):
+        retrying_wrapper = self.retry(unwrapped)
+
+        @wraps(unwrapped)
+        def deciding_wrapper(aws_retry=False, *args, **kwargs):
+            if aws_retry:
+                return retrying_wrapper(*args, **kwargs)
+            else:
+                return unwrapped(*args, **kwargs)
+        return deciding_wrapper
+
+    def __getattr__(self, name):
+        unwrapped = getattr(self.client, name)
+        if name in self.__never_wait:
+            return unwrapped
+        elif callable(unwrapped):
+            wrapped = self._create_optional_retry_wrapper_function(unwrapped)
+            setattr(self, name, wrapped)
+            return wrapped
+        else:
+            return unwrapped
